Fix the demo
