+++ Mesh_3/include/CGAL/Polyhedral_mesh_domain_with_features_3.h
@@ -40,7 +40,7 @@
 #include <CGAL/Mesh_3/Detect_polylines_in_polyhedra.h>
 #include <CGAL/Mesh_3/Polyline_with_context.h>
 #include <CGAL/Polygon_mesh_processing/Detect_features_in_polyhedra.h>
-#include <CGAL/boost/graph/PMP_properties_Polyhedron_3.h>
+#include <CGAL/Mesh_3/properties_Polyhedron_3.h>
 
 #include <CGAL/enum.h>
 #include <CGAL/IO/Polyhedron_iostream.h>
