+++ Polyhedron/demo/Polyhedron/Plugins/Mesh_3/Mesh_3_plugin_cgal_code.h
@@ -52,6 +52,7 @@ Meshing_thread* cgal_code_mesh_3(const SMwgd* pMesh,
                                  const double edge_size,
                                  const double tet_shape,
                                  bool protect_features,
+                                 bool protect_border,
                                  const double sharp_edges_angle,
                                  const int manifold,
                                  const bool surface_only,
